Projectize Spring Config
