+++ web-rest/build.gradle
@@ -8,6 +8,7 @@ plugins {
 version = '0.0.1-SNAPSHOT'
 
 dependencies {
+    implementation 'org.springframework:spring-webmvc' 
     //implementation 'org.springframework.boot:spring-boot-starter-web'
     implementation('org.springframework.boot:spring-boot-starter-webflux')
     implementation 'org.springframework.boot:spring-boot-starter-actuator'
